Rename `writeback` directory to `data-apps`
