+++ frontend/src/metabase-types/api/database.ts
@@ -21,7 +21,6 @@ export interface Database {
   native_permissions: NativePermissions;
   initial_sync_status: InitialSyncStatus;
 
-  // appears in frontend/src/metabase/writeback/utils.ts
   settings?: DatabaseSettings | null;
 
   // Only appears in  GET /api/database/:id
