+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/TypeSelector/TypeSelector.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useMemo } from "react";
 import Icon from "metabase/components/Icon";
 import { color } from "metabase/lib/colors";
 
-import { isActionDashCard } from "metabase/writeback/utils";
+import { isActionDashCard } from "metabase/data-apps/utils";
 
 import type { DashboardOrderedCard, ClickBehavior } from "metabase-types/api";
 import type { UiParameter } from "metabase-lib/parameters/types";
