+++ frontend/src/metabase/data-apps/components/DataAppPageTitle/Suggestions.tsx
