+++ frontend/src/metabase/data-apps/containers/CreateDataAppModal/DataAppScaffoldingModal.tsx
@@ -20,8 +20,8 @@ import DataPicker, {
 import {
   scaffoldDataApp,
   ScaffoldNewAppParams,
-} from "metabase/writeback/actions";
-import DataAppScaffoldingDataPicker from "metabase/writeback/components/DataAppScaffoldingDataPicker";
+} from "metabase/data-apps/actions";
+import DataAppScaffoldingDataPicker from "metabase/data-apps/components/DataAppScaffoldingDataPicker";
 
 import type { DataApp } from "metabase-types/api";
 import type { Dispatch, State } from "metabase-types/store";
